switched implementation of read-file to use file_read_raw

+++ worker/file.c
@@ -18,13 +18,13 @@ Bool file_read_raw(char *filename, char **buffer, size_t *size) {
         return false;
     }
     fseek(file, 0, SEEK_END);
-    long file_size = ftell(file);
+    size_t file_size = ftell(file);
     w_assert(file_size > 0);
     rewind(file);
 
     char *allocated = NEW_BUFFER(char, file_size + 1);
     size_t bytes_read = fread(allocated, sizeof(char), file_size, file);
-    if (bytes_read != (size_t) file_size) {
+    if (bytes_read != file_size) {
         log_error("Could not read entire file %s", filename);
         fclose(file);
         memory_free(allocated);
